split form into seperate fields
