+++ modules/costs/app/components/time_entries/days_and_hours_form.rb
@@ -1,0 +1,89 @@
+# frozen_string_literal: true
+
+#-- copyright
+# OpenProject is an open source project management software.
+# Copyright (C) the OpenProject GmbH
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License version 3.
+#
+# OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
+# Copyright (C) 2006-2013 Jean-Philippe Lang
+# Copyright (C) 2010-2013 the ChiliProject Team
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+#
+# See COPYRIGHT and LICENSE files for more details.
+#++
+
+module TimeEntries
+  class DaysAndHoursForm < ApplicationForm
+    delegate :project, to: :model
+
+    form do |f|
+      f.single_date_picker name: :spent_on,
+                           type: "date",
+                           required: true,
+                           datepicker_options: { inDialog: true },
+                           value: model.spent_on&.iso8601,
+                           label: TimeEntry.human_attribute_name(:spent_on)
+
+      if show_start_and_end_time_fields?
+        f.group(layout: :horizontal) do |g|
+          g.text_field name: :start_time,
+                       type: "time",
+                       required: true,
+                       label: TimeEntry.human_attribute_name(:start_time),
+                       value: model.start_timestamp&.strftime("%H:%M"),
+                       data: {
+                         "time-entry-target" => "startTimeInput",
+                         "action" => "input->time-entry#timeInputChanged"
+                       }
+
+          g.text_field name: :end_time,
+                       type: "time",
+                       required: true,
+                       label: TimeEntry.human_attribute_name(:end_time),
+                       value: model.end_timestamp&.strftime("%H:%M"),
+                       data: {
+                         "time-entry-target" => "endTimeInput",
+                         "action" => "input->time-entry#timeInputChanged"
+                       }
+        end
+      end
+
+      f.text_field name: :hours,
+                   required: true,
+                   label: TimeEntry.human_attribute_name(:hours),
+                   value: hours_value,
+                   data: { "time-entry-target" => "hoursInput",
+                           "action" => "blur->time-entry#hoursChanged keypress.enter->time-entry#hoursKeyEnterPress" }
+    end
+
+    private
+
+    def show_start_and_end_time_fields?
+      TimeEntry.can_track_start_and_end_time?
+    end
+
+    def hours_value
+      if model.hours.present?
+        ChronicDuration.output(model.hours * 3600, format: :hours_only)
+      else
+        ""
+      end
+    end
+  end
+end
