+++ modules/costs/app/components/time_entries/time_entry_form_component.html.erb
@@ -23,14 +23,19 @@ You should have received a copy of the GNU General Public License
 along with this program; if not, write to the Free Software
 Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
-See COPYRIGHT and LICENSE files for more details.
+  See COPYRIGHT and LICENSE files for more details.
 
-++#%>
+  ++# %>
 
-<%=
-  component_wrapper do
+  <%= component_wrapper do
     primer_form_with(**form_options) do |form|
-            render(TimeEntries::TimeEntryForm.new(form))
-        end
+      flex_layout do |body|
+        body.with_row(mb: 3) { render(TimeEntries::UserForm.new(form)) }
+        body.with_row(mb: 3) { render(TimeEntries::DaysAndHoursForm.new(form)) }
+        body.with_row(mb: 3) { render(TimeEntries::WorkPackageForm.new(form)) }
+        body.with_row(mb: 3) { render(TimeEntries::ActivityForm.new(form)) }
+        body.with_row(mb: 3) { render(TimeEntries::CommentsForm.new(form)) }
+        body.with_row(mb: 3) { render(TimeEntries::CustomFieldsForm.new(form)) }
       end
-%>
+    end
+  end %>
