+++ modules/costs/app/components/time_entries/work_package_form.rb
@@ -1,0 +1,56 @@
+# frozen_string_literal: true
+
+#-- copyright
+# OpenProject is an open source project management software.
+# Copyright (C) the OpenProject GmbH
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License version 3.
+#
+# OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
+# Copyright (C) 2006-2013 Jean-Philippe Lang
+# Copyright (C) 2010-2013 the ChiliProject Team
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+#
+# See COPYRIGHT and LICENSE files for more details.
+#++
+
+module TimeEntries
+  class WorkPackageForm < ApplicationForm
+    form do |f|
+      f.work_package_autocompleter name: :work_package_id,
+                                   label: TimeEntry.human_attribute_name(:work_package),
+                                   required: true,
+                                   autocomplete_options: {
+                                     focusDirectly: false,
+                                     append_to: "#time-entry-dialog",
+                                     filters: work_package_completer_filters
+                                   }
+    end
+
+    private
+
+    def work_package_completer_filters
+      filters = []
+
+      if model.project_id
+        filters << { name: "project_id", operator: "=", values: [model.project_id] }
+      end
+
+      filters
+    end
+  end
+end
